Updated Enemy includes spawner
Enemy spawner game object with the script attached to handle the spawning ⁠Debugged the SpatulaCol script to destroy enemies on collision. Destroys spatulas that collide with the enemy. Added code to destroy remaining spatulas in the scene as some were remaining. Added some buzzing audio to the enemies as well (if it’s not necessary it can be removed). Also on spatula hit.